- added ProjectConfigurationBuilder.cs
