+++ src/Nox.Core/Nox.Core.csproj
@@ -20,9 +20,9 @@
     <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
     <Copyright>Copyright (c) Andre Sharpe 2022</Copyright>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <AssemblyVersion>6.0.47.0</AssemblyVersion>
-    <FileVersion>6.0.47.0</FileVersion>
-    <PackageVersion>6.0.47</PackageVersion>
+    <AssemblyVersion>6.0.48.0</AssemblyVersion>
+    <FileVersion>6.0.48.0</FileVersion>
+    <PackageVersion>6.0.48</PackageVersion>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
     <PackageProjectUrl>https://github.com/NoxOrg/Nox</PackageProjectUrl>
     <RepositoryUrl>https://github.com/NoxOrg/Nox.git</RepositoryUrl>
