+++ src/Nox.Data.Csv/Nox.Data.Csv.csproj
@@ -6,8 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
   <PropertyGroup>
-    <AssemblyVersion>6.0.47.0</AssemblyVersion>
-    <FileVersion>6.0.47.0</FileVersion>
+    <AssemblyVersion>6.0.48.0</AssemblyVersion>
+    <FileVersion>6.0.48.0</FileVersion>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
     <DebugType>none</DebugType>
